The HUD is working!
I got the HUD for the game working. You can now reset, and change
character with the next botton. The HUDs code also keeps track of
managing the current character. You can now throw three characters in
any order before the change character button does not work any more.
This means the user has used all their throws, and we can use that
variable to know we need to check if the level is complete.